/project/explore/{project id} /bids - Manage Bidders - Make Bidding - My Active Bidding - Remove Bidding

+++ app/Http/Controllers/ProjectController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Bidding;
 use App\Models\Project;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
@@ -63,13 +65,25 @@ class ProjectController extends Controller
 
     public function show($id)
     {
+        if (Auth::user()) {
+            $Project = Project::find($id);
+            $deadline = Carbon::parse($Project->deadline);
+            return view('project.project-bid', compact('Project', 'deadline'));
+        } else {
+            return redirect('/login');
+        }
 
     }
 
     public function edit($id)
     {
-        $Project = Project::find($id);
-        return view('project.project-edit', compact('Project'));
+        if (Auth::user()) {
+            $Project = Project::find($id);
+            return view('project.project-edit', compact('Project'));
+        } else {
+            return redirect('/login');
+        }
+
     }
 
     public function update(Request $request, $id)
@@ -100,16 +114,76 @@ class ProjectController extends Controller
 
     public function destroy($id, Request $request)
     {
-        $Project = Project::find($id);
-        $Project->delete();
-        return redirect('/projects')->with('success', 'Project deleted successfully');
+        if (Auth::user()) {
+            $Project = Project::find($id);
+            $Project->delete();
+            return redirect('/projects')->with('success', 'Project deleted successfully');
+        } else {
+            return redirect('/login');
+        }
+
     }
 
     public function bid($id)
     {
-        $Project = Project::find($id);
+        if (Auth::user()) {
+            $Project = Project::find($id);
 //        $Project->bids(1);
 //        dd($Project);
-        return view('project.manage-bidders', compact('Project'));
+            return view('project.manage-bidders', compact('Project'));
+        } else {
+            return redirect('/login');
+        }
+
     }
+
+    public function bidding(Request $request)
+    {
+        if (Auth::user()) {
+            $validator = Validator::make($request->all(), [
+                'price' => 'required|numeric',
+                'days' => 'required|numeric',
+                'notes' => 'required|string',
+            ]);
+            if ($validator->fails())
+                return redirect()->back()->WithErrors($validator->errors()->all())->withInput($request->except('password'));
+            else {
+                $data = $request->all();
+                $data['id_user_designer'] = $request->user()->id;
+                $data['id_user_project_owner'] = $request->id_project;
+//        dd($data);
+                Bidding::create($data);
+                return redirect()->back()->with('success', 'Project Published successfully.');
+            }
+
+        } else {
+            return redirect('/login');
+        }
+    }
+
+    public function bids(Request $request)
+    {
+
+        if (Auth::user()) {
+            $Bidding = Bidding::where('id_user_designer', $request->user()->id)->get();
+            return view('project.show-bids', compact('Bidding'));
+
+
+        } else {
+            return redirect('/login');
+        }
+    }
+
+    public function destroyBid($id, Request $request)
+    {
+        if (Auth::user()) {
+            $Bidding = Bidding::find($id);
+            $Bidding->delete();
+            return redirect('/bids')->with('success', 'Bid deleted successfully');
+        } else {
+            return redirect('/login');
+        }
+    }
+
+
 }
